Search works, Some manufacturing items are populated, Main items and group menu items are loaded in parallel (yay speed boost)

+++ EvE Build WPF/Code/Containers/Types.cs
@@ -38,9 +38,9 @@
     public class Cost
     {
         public int StationId { get; private set; }
-        public long Price { get; private set; }
+        public decimal Price { get; private set; }
 
-        public Cost(int stationId, long price)
+        public Cost(int stationId, decimal price)
         {
             StationId = stationId;
             Price = price;
